requirements.txt
bs4
Pillow>=5.3.0
python-dateutil
pycountry>=18.12.8
requests